rename: 학과 정보 파일명 수정
